Move jiffy functions from cspad_detector_congruence to serialtbx.detector
